refactor: Car,Cars,CarLocation 객체를 vo 패키지로 이동 및 CarLocation 클래스에 toString() 함수 오버라이딩 추가
