+++ src/main/java/racinggame/model/CarMover.java
@@ -3,6 +3,7 @@ package racinggame.model;
 import nextstep.utils.Randoms;
 import racinggame.enums.RANDOM_MOVE_RESULT;
 import racinggame.environments.GlobalVariables;
+import racinggame.vo.Car;
 
 public class CarMover {
     public static RANDOM_MOVE_RESULT moveRandomizly(Car car) {
