--- src/main/java/racinggame/model/GameTurnResult.java
@@ -1,8 +1,0 @@
-package racinggame.model;
-
-import racinggame.enums.RANDOM_MOVE_RESULT;
-
-import java.util.ArrayList;
-
-public class GameTurnResult extends ArrayList<RANDOM_MOVE_RESULT> {
-}
