--- src/main/java/racinggame/model/GameTurnResults.java
@@ -1,6 +1,0 @@
-package racinggame.model;
-
-import java.util.ArrayList;
-
-public class GameTurnResults extends ArrayList<GameTurnResult> {
-}
