+++ src/main/java/racinggame/vo/Cars.java
@@ -1,4 +1,4 @@
-package racinggame.model;
+package racinggame.vo;
 
 import java.util.ArrayList;
