+++ src/test/java/racinggame/model/CarTest.java
@@ -3,6 +3,8 @@ package racinggame.model;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racinggame.vo.Car;
+import racinggame.vo.CarLocation;
 import racinggame.vo.PlayerName;
 
 public class CarTest {
